vertical_1_fin-anlyzer/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
pdf2image==1.17.0
Pillow==10.3.0
python-multipart==0.0.9